weather_fetcher_mcp_server_client/requirements.txt
mcp-sdk
requests